move api folder to src
