+++ packages/app-core/src/hooks/use-ohlc.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback, useEffectEvent } from 'react';
 import type { PublicRestTypes } from 'ts-kraken';
-import { fetchOHLC, type OHLCInterval } from '../services/api/kraken-rest-api';
+import { fetchOHLC, type OHLCInterval } from '../api/kraken-rest-api';
 import { mapOHLCResponse, mergeCandles } from '../mappers/candle-mapper';
 import type { Candle } from '../domain/Candle';
