+++ packages/app-core/src/services/business/auth-service.ts
@@ -7,9 +7,9 @@
  */
 
 import { AuthSession } from '../../domain/AuthSession';
-import * as authApi from '../api/auth-api';
+import * as authApi from '../../api/auth-api';
 import { useAuthStore } from '../../state/auth-store';
-import { resetWsToken } from '../api/kraken-ws-api';
+import { resetWsToken } from '../../api/kraken-ws-api';
 
 const AUTH_STORAGE_KEY = 'kraken_auth_session';
